Create CalenderReminder
This is my first project in which I made a Calendar reminder . On my first day I made 3 menu bars, The first one is adding a reminder, the second one is viewing the reminder and the third one is to exit the menu bar. And in the furthur days i will do some add-on in this program.